Create README.md
Twitter Clone is a small project that feature logged in users, a customized homepage, a tweet composing page, the ability to follow other users, and more. 
This project is divided into two parts. Part 1 is all about the folder structure. Part 2 is going to be implementing the logic and front end.

There are four pieces to this application, each with their own app:

*   twitterclone <-- This is your project folder
*   twitteruser <-- This is a custom user app!
*   tweet
*   authentication
*   notification